Create Temp_C_code.c
This code is temporary code that is used to test sending and receiving data from Raspberry Pi. There is one issue when receiving data from the Pi and that is the first byte is always not correct. I worked around that issue by sending a byte that can be disregarded and on the C code side I read starting from the second byte. So far the test has been done only with a string. Next step is to test it with bytes of the same type as the Pi would originally send and check the correctness of the received bytes.